Enemy
Enemy was added

+++ Assets/Scripts/AI/UI/IAIUIView.cs
@@ -5,6 +5,6 @@
         int MoneySet { set; }
         int HealthSet { set; }
         int ForceSet { set; }
-        int EnemyPowerSet { set; }
+        float EnemyPowerSet { set; }
     }
 }
